+++ tests/run_tests.ps1
@@ -68,34 +68,34 @@ run_hook -Hook cmake-pc-cppcheck-hook -Src tests/cmake_bad -ExpectSuccess 0 -Gen
 
 # ==============================================================================
 
-# Commented out since Include-What-You-Use is currently not installable on Windows
-# (at least that I know of)
-# cd tests/cmake_good
-# git init
-# git config user.name 'Test'
-# git config user.email 'test@test.com'
-# git add *.txt *.cpp .pre-commit-config.yaml
-# git commit -m 'Initial commit'
-# pre-commit run --all-files
-# rm -Force -Recurse .git
-# if ($? -eq 0) {
-#     echo "Pre-commit on tests/cmake_good failed!"
-#     Exit 1
-# }
-# cd ..
+cd tests/cmake_good
+git init
+git config user.name 'Test'
+git config user.email 'test@test.com'
+git add *.txt *.cpp .pre-commit*.yaml
+git commit -m 'Initial commit'
+pre-commit run -c .pre-commit-win.yaml --all-files
+$ret=$?
+rm -Force -Recurse .git
+if ($ret -eq 0) {
+    echo "Pre-commit on tests/cmake_good failed!"
+    Exit 1
+}
+cd ../..
 
-# cd tests/cmake_bad
-# git init
-# git config user.name 'Test'
-# git config user.email 'test@test.com'
-# git add *.txt *.cpp .pre-commit-config.yaml
-# git commit -m 'Initial commit'
-# pre-commit run --all-files
-# rm -Force -Recurse .git
-# if ($? -eq 1) {
-#     echo "Pre-commit on tests/cmake_good unexpectedly passed!"
-#     Exit 1
-# }
-# cd ..
+cd tests/cmake_bad
+git init
+git config user.name 'Test'
+git config user.email 'test@test.com'
+git add *.txt *.cpp .pre-commit*.yaml
+git commit -m 'Initial commit'
+pre-commit run -c .pre-commit-win.yaml --all-files
+$ret=$?
+rm -Force -Recurse .git
+if ($ret -eq 1) {
+    echo "Pre-commit on tests/cmake_bad unexpectedly passed!"
+    Exit 1
+}
+cd ../..
 
 Exit 0